Flipped canvas
Cameras flip your selfies because it gives you the image of what you look like to other people. This is the flipped version of what you see in your reflection and people do not usually see themselves in that way. After you take the picture you are looking at yourself from an outsider’s perspective.